refactor: move pickAppleNotesFolder to utils/prompts/clack
