+++ .gitignore
@@ -199,3 +199,6 @@ vercel-skills
 
 # HAR network captures
 *.har
+
+.claude/worktrees/
+.worktrees/
