+++ src/utils/prompts/clack/index.ts
@@ -9,3 +9,4 @@ export * from "./multiline";
 export * from "./search-multiselect";
 export * from "./search-select";
 export * from "./settings-summary";
+export { pickAppleNotesFolder } from "./apple-notes";
